[ACS-5742] added People filter to demo-shell
